Revert "merged task2 into task4"
This reverts commit dac3a944c67cf20426fbb044a99eafc95b67d075, reversing
changes made to 5abc3cda9951253ba221a502c5ff15c6ec39c528.

+++ src/particle/container/ParticleContainerDirectSum.cpp
@@ -27,7 +27,7 @@ std::unique_ptr<ParticleContainer> ParticleContainerDirectSum::copy() {
     return std::make_unique<ParticleContainerDirectSum>(*this);
 }
 
-Particle &ParticleContainerDirectSum::getParticle(size_t index) {
+Particle &ParticleContainerDirectSum::getParticle(int index) {
     return particles.at(index);
 }
